chore(valueFile): move onHost code to specific folder
