+++ super-agent/src/opamp/hash_repository/mod.rs
@@ -1,7 +1,6 @@
 pub mod repository;
 pub use repository::HashRepository;
 
-// TODO to be moved below onhost cfg flag when k8s implementation is ready.
 #[cfg(all(not(feature = "onhost"), feature = "k8s"))]
 mod k8s;
 #[cfg(feature = "onhost")]
