+++ super-agent/test/on_host/values_repository.rs
@@ -1,15 +1,13 @@
+use ::fs::directory_manager::{DirectoryManager, DirectoryManagerFs};
+use newrelic_super_agent::agent_type::agent_values::AgentValues;
+use newrelic_super_agent::sub_agent::values::values_repository::ValuesRepository;
+use newrelic_super_agent::sub_agent::values::ValuesRepositoryFile;
+use newrelic_super_agent::super_agent::config::AgentID;
 use std::fs;
 use std::fs::Permissions;
 #[cfg(target_family = "unix")]
 use std::os::unix::fs::PermissionsExt;
 use std::path::PathBuf;
-
-use ::fs::directory_manager::{DirectoryManager, DirectoryManagerFs};
-use newrelic_super_agent::agent_type::agent_values::AgentValues;
-use newrelic_super_agent::sub_agent::values::values_repository::{
-    ValuesRepository, ValuesRepositoryFile,
-};
-use newrelic_super_agent::super_agent::config::AgentID;
 
 // This test is the only one that writes to an actual file in the FS
 #[test]
